--- common/views/components/Tags/Tags.config.js
@@ -1,16 +1,0 @@
-export const status = 'graduated';
-export const context = {
-  tags: [
-    {
-      text: 'Event',
-      url: '#',
-    },
-    {
-      text: 'Walking tour',
-    },
-    {
-      text: 'A Museum of Modern Nature',
-      url: '/components/preview/exhibition-template.html',
-    },
-  ],
-};
