--- common/views/components/WorkTags/README.md
@@ -1,3 +1,0 @@
-## Purpose
-
-To indicate the links on work pages that will trigger a search.
